Rename redux folder to services
