+++ src/components/order-details/order-details.tsx
@@ -4,7 +4,7 @@ import { useSelector } from 'react-redux';
 import { CheckMarkIcon } from '@ya.praktikum/react-developer-burger-ui-components';
 import Loader from '../ui/loader/loader';
 import styles from './order-details.module.css';
-import { RootState } from '../../redux/store';
+import { RootState } from '../../services/store';
 
 export default function OrderDetails() {
   const order = useSelector((state: RootState) => state.order.order);
